Use proper param construction for search Rename validate.py to validator.py Remove unneeded Dockerfile-dev (better caching)
