+++ namer/validator.py
